% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_various_random_factor_models.R
\name{EvaluateVariousFactorModelsR}
\alias{EvaluateVariousFactorModelsR}
\title{Evaluate}
\usage{
EvaluateVariousFactorModelsR(
  returns,
  factors,
  penalty_parameters,
  n_evaluated_models = 10000,
  n_kept_factors = 0,
  weighting_type = "c",
  one_stddev_rule = TRUE,
  gcv_vr_weighting = FALSE,
  gcv_scaling_n_assets = FALSE,
  gcv_identification_check = FALSE,
  n_bootstrap_cf2019_rank_test = 500,
  target_level_kp2006_rank_test = 0.05,
  hj_test_level = 0.05,
  rng_seed = 1
)
}
\arguments{
\item{returns}{returns}

\item{factors}{all_factors}

\item{penalty_parameters}{penalty_parameters}

\item{n_evaluated_models}{n_models}

\item{n_kept_factors}{n_kept_factors}

\item{weighting_type}{character specifying the type of adaptive weights:
based on the correlation between factors and returns 'c'; based on the
regression coefficients of returns on factors 'b'; based on the first-step
intrinsic risk premia estimator 'a'; otherwise a vector of ones (any other
character). Default is 'c'.}

\item{one_stddev_rule}{boolean TRUE for picking the most parsimonious model
whose score is not higher than one standard error above the score of the
best model; FALSE for picking the best model. Default is TRUE.}

\item{gcv_vr_weighting}{boolean \code{TRUE} for scaling pricing errors by
the inverse variance matrix of asset excess returns; \code{FALSE} otherwise.
Default is \code{FALSE}.}

\item{gcv_scaling_n_assets}{(only relevant for \code{tuning_type ='g'})
boolean \code{TRUE} for log(n_assets) scaling (\code{log(n_assets) / n_observations});
\code{FALSE} otherwise (\code{1 / n_observations}). Default is \code{FALSE}.}

\item{gcv_identification_check}{(only relevant for \code{tuning_type ='g'})
boolean \code{TRUE} for a loose check for model identification; \code{FALSE} otherwise.
Default is \code{FALSE}.}

\item{n_bootstrap_cf2019_rank_test}{(only relevant if
\code{gcv_identification_check} is \code{TRUE}) number of bootstrap samples in the
Chen Fang 2019 rank test. Default is \code{500}.}

\item{target_level_kp2006_rank_test}{(only relevant if \code{gcv_identification_check} is
\code{TRUE}) numeric level of the Kleibergen Paap 2006 rank test. If it is
strictly grater than zero, then the iterative Kleibergen Paap 2006 rank
test at \code{level = target_level_kp2006_rank_test / n_factors} is used to compute an initial estimator
of the rank of the factor loadings in the Chen Fang 2019 rank test.
Otherwise, the initial rank estimator is taken to be the number of singular
values above \code{n_observations^(-1/4)}. Default is \code{0.05} (as correction
for multiple testing).}

\item{hj_test_level}{numeric level of the HJ misspecification test.
Default is \code{0.05}.}

\item{rng_seed}{random number generation seed}
}
\value{
a list containing the values of the HJ-misspecification statistics
and the beta rank statistics
}
\description{
\code{EvaluateVariousFactorModelsR}
}
