\name{reproduceTFRP-package}
\alias{reproduceTFRP-package}
\alias{reproduceTFRP}
\docType{package}
\title{reproduceTFRP: Reproduction Code for Tradable Factor Risk Premia}
\description{
Minimal package scaffold for organizing data, code, and replication
workflows for tradable factor risk premia research.
}
\keyword{internal}
